Heida frontend

+++ assets/scripts/controllers/report.js
@@ -43,13 +43,18 @@ angular.module('heidaApp')
   }
 
     $scope.deleteReport = function(dataId) {
-      $http.delete('/api/data/' + dataId.dataId)
-        .success(function () {
-          $state.go('dashboard.report', $stateParams, {
-            reload: true,
-            inherit: true
-          });
-        })
+      if (confirm('Are you sure you want to delete this indicator?')) {
+        $http.delete('/api/data/' + dataId.dataId)
+          .success(function () {
+            $state.go('dashboard.report', $stateParams, {
+              reload: true,
+              inherit: true
+            });
+          })
+      } else {
+        // Do nothing!
+      }
+
     }
 
 })
@@ -84,7 +89,7 @@ angular.module('heidaApp')
   if(printReport[2] == 'print') {
     setTimeout(function(){
       window.print();
-    }, 300);
+    }, 1000);
   }
 });
